fix: move otel setup to separated file
