+++ packages/app/example.env
@@ -33,6 +33,9 @@ SWAGGER_SERVER=/
 # Axios request timeout, default "60000"
 HTTP_DEFAULT_CLIENT_TIMEOUT=60000
 
+# Service name for Open Telemetry
+SERVICE_NAME=<App>
+
 ################################################################################
 
 # Logger level, default "info"
